requirements.txt
pandas
numpy
scikit-learn
scikit-surprise
flask
joblib
gunicorn
requests